サインインとチーム振り分けに関する関数の切り分け完了
