+++ src/app/page.tsx
@@ -1,11 +1,20 @@
 "use client";
 
+import TeamButtons from "@/components/TeamButtons";
 import TeamCard from "@/components/TeamCard";
+import { fetchTeamsCount } from "@/logics/CountTheNumberOfTeams";
+import { signInWithAnonymous } from "@/logics/SignInWithAnonymous";
+import { auth, db } from "@/logics/firebase";
 import { Team } from "@/logics/types/team";
+import { onAuthStateChanged, signInAnonymously } from "firebase/auth";
+import { collection, getDocs } from "firebase/firestore";
 import { useRouter } from "next/navigation";
-import { useState } from "react";
+import { useEffect, useState } from "react";
+import { useAuthState } from "react-firebase-hooks/auth";
 
 export default function Home() {
+  const [user] = useAuthState(auth);
+  const [teamsCount, setTeamsCount] = useState(0);
 
   const [teams, setTeams] = useState<Team[]>([
     { name: "Team1", id: "1" },
@@ -18,17 +27,39 @@ export default function Home() {
 
   const teamCards = teams.map((team) => <TeamCard key={team.id} router={router}>{team.name}</TeamCard>);
 
+  useEffect(() => {
+    const getTeamsCount = async () => {
+      const count = await fetchTeamsCount();
+      if (count != undefined) {
+        setTeamsCount(count);
+      }
+    }
+    getTeamsCount();
+  }, [])
+
+  // function gameStartButtonHandler() {
+  //   signInWithAnonymous
+  // }
+
   return (
-    <div className="text-center">
-      <div>トロッコ問題ゲーム！</div>
-      <div>
-        <div>チーム一覧</div>
-        <div className="grid auto-rows-[6rem] grid-cols-[repeat(auto-fill,minmax(8rem,1fr))] gap-2 overflow-y-auto p-2">
-          {teamCards}
+    <div>
+      {user ? (
+        <div className="text-center">
+          チームをセレクト
+          <div>
+            <TeamButtons teamsCount={teamsCount} />
+          </div>
         </div>
-        <div>開始までお待ち下さい…………</div>
-        <button onClick={() => setTeams([...teams, { name: "Team5", id: "5" }])}>addTeams</button>
-      </div>
+
+      ) : (
+        <div className="text-center">
+          <div>管理画面</div>
+          <button
+            onClick={signInWithAnonymous}>
+            ゲームを開始
+          </button>
+        </div>
+      )}
     </div>
   );
 }
